Change the folder of Chen
