+++ Content/Items/Armor/Vanity/Guard/ChenBody.png
