+++ Content/Items/Armor/Vanity/Guard/ChenBody_Body.png
