+++ Content/Items/Armor/Vanity/Guard/ChenHead.png
