+++ Content/Items/Armor/Vanity/Guard/ChenHead_Head.png
